LPS-60606 Move ConfigurationPortletConfigurationIcon classes to PortletConfigurationPortlet and removed addon since it's no longer used
