--- modules/addons/portlet-configuration-icon/portlet-configuration-icon-configuration/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Portlet Configuration Icon Configuration
-Bundle-SymbolicName: com.liferay.portlet.configuration.icon.configuration
-Bundle-Version: 1.0.0
-Include-Resource: classes
