--- modules/addons/portlet-configuration-icon/portlet-configuration-icon-configuration/build.gradle
@@ -1,9 +1,0 @@
-dependencies {
-	compile group: "com.liferay.portal", name: "portal-service", version: liferay.portalVersion
-	compile group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
-	compile group: "org.osgi", name: "org.osgi.compendium", version: "5.0.0"
-}
-
-liferay {
-	deployDir = file("${liferayHome}/osgi/modules")
-}
