--- modules/addons/portlet-configuration-icon/portlet-configuration-icon-configuration/build.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<!DOCTYPE project>
-
-<project>
-	<import file="../../../build-module.xml" />
-</project>
